Added validator for Product model

+++ GroceryMarket/ModelsDTO/ProductDto.cs
@@ -5,7 +5,7 @@
         public string ProductCode { get; set; }
         public double Price { get; set; }
         public double WholesalePrice { get; set; }
-        public double WholesaleCount { get; set; }
+        public int WholesaleCount { get; set; }
         public bool IsWholesale { get; set; }
     }
 }
